+++ docs/src/components/codeEditorControls/FindReplace.js
@@ -1,7 +1,3 @@
-import e from"./CodeEditorControl.js";import{html as t,css as i}from"../../lit-all.min.js";import"../Icon.js";export default class s extends e{static styles=[e.styles,i`
+import e from"./CodeEditorButtonControl.js";import{html as t,css as i}from"../../lit-all.min.js";import"../Icon.js";export default class n extends e{static styles=[e.styles,i`
 			:host { display: inline-flex; }
-		`];handleClick=()=>{this.editor?.openFind()};render(){return t`
-			<button class="${this.buttonClasses}" @click="${this.handleClick}" title="Find & Replace">
-				<k-icon name="search"></k-icon>
-			</button>
-		`}}customElements.define("k-cec-find-replace",s);
+		`];connectedCallback(){super.connectedCallback(),this.hasAttribute("title")||(this.title="Find & Replace")}handleAction(){this.editor?.openFind()}render(){return t`<k-icon name="search"></k-icon>`}}customElements.define("k-cec-find-replace",n);