+++ docs/src/components/codeEditorControls/FontSize.js
@@ -1,13 +1,11 @@
-import e from"./CodeEditorControl.js";import{html as t,css as s}from"../../lit-all.min.js";import"../Icon.js";export default class n extends e{static styles=[e.styles,s`
+import e from"./CodeEditorControl.js";import{html as s,css as t}from"../../lit-all.min.js";import"./FontSizeDecrease.js";import"./FontSizeIncrease.js";import"./ControlGroup.js";export default class o extends e{static styles=[e.styles,t`
 			:host {
 				display: inline-flex;
 				gap: 0;
 			}
-		`];handleIncrease=()=>{this.editor?.increaseFontSize()};handleDecrease=()=>{this.editor?.decreaseFontSize()};render(){return t`
-			<button class="${this.buttonClasses}" @click="${this.handleDecrease}" title="Decrease Font Size">
-				<k-icon name="text_decrease"></k-icon>
-			</button>
-			<button class="${this.buttonClasses}" @click="${this.handleIncrease}" title="Increase Font Size">
-				<k-icon name="text_increase"></k-icon>
-			</button>
-		`}}customElements.define("k-cec-font-size",n);
+		`];render(){return s`
+			<k-cec-group class="b r mq">
+				<k-cec-font-size-decrease></k-cec-font-size-decrease>
+				<k-cec-font-size-increase></k-cec-font-size-increase>
+			</k-cec-group>
+		`}}customElements.define("k-cec-font-size",o);