+++ src/components/codeEditorControls/FindReplace.js
@@ -1,34 +1,38 @@
-import CodeEditorControl from './CodeEditorControl.js';
+import CodeEditorButtonControl from './CodeEditorButtonControl.js';
 import { html, css } from '../../lit-all.min.js';
 import '../Icon.js';
 
-export default class FindReplace extends CodeEditorControl {
+export default class FindReplace extends CodeEditorButtonControl {
 	/*
 		Styles
 	*/
 	static styles = [
-		CodeEditorControl.styles,
+		CodeEditorButtonControl.styles,
 		css`
 			:host { display: inline-flex; }
 		`
 	];
 
 	/*
-		Event Handlers
+		Lifecycle Callbacks
 	*/
-	handleClick = () => {
+	connectedCallback() {
+		super.connectedCallback();
+		if(!this.hasAttribute('title')) this.title = 'Find & Replace';
+	}
+
+	/*
+		Public Methods
+	*/
+	handleAction() {
 		this.editor?.openFind();
-	};
+	}
 
 	/*
 		Rendering
 	*/
 	render() {
-		return html`
-			<button class="${this.buttonClasses}" @click="${this.handleClick}" title="Find & Replace">
-				<k-icon name="search"></k-icon>
-			</button>
-		`;
+		return html`<k-icon name="search"></k-icon>`;
 	}
 }