+++ src/components/codeEditorControls/Undo.js
@@ -1,34 +1,38 @@
-import CodeEditorControl from './CodeEditorControl.js';
+import CodeEditorButtonControl from './CodeEditorButtonControl.js';
 import { html, css } from '../../lit-all.min.js';
 import '../Icon.js';
 
-export default class Undo extends CodeEditorControl {
+export default class Undo extends CodeEditorButtonControl {
 	/*
 		Styles
 	*/
 	static styles = [
-		CodeEditorControl.styles,
+		CodeEditorButtonControl.styles,
 		css`
 			:host { display: inline-flex; }
 		`
 	];
 
 	/*
-		Event Handlers
+		Lifecycle Callbacks
 	*/
-	handleClick = () => {
+	connectedCallback() {
+		super.connectedCallback();
+		if(!this.hasAttribute('title')) this.title = 'Undo';
+	}
+
+	/*
+		Public Methods
+	*/
+	handleAction() {
 		this.editor?.undo();
-	};
+	}
 
 	/*
 		Rendering
 	*/
 	render() {
-		return html`
-			<button class="${this.buttonClasses}" @click="${this.handleClick}" title="Undo">
-				<k-icon name="undo"></k-icon>
-			</button>
-		`;
+		return html`<k-icon name="undo"></k-icon>`;
 	}
 }