+++ tests/components/codeEditorControls/CopyCode.browser-test.js
@@ -20,12 +20,12 @@ const cleanup = container => {
 };
 
 export default {
-	'should create element with button': async ({pass, fail}) => {
+	'should have button role': async ({pass, fail}) => {
 		const { container, control } = await createEditorWithControl();
 		await control.updateComplete;
-		if(!control.shadowRoot.querySelector('button')){
+		if(control.getAttribute('role') !== 'button'){
 			cleanup(container);
-			return fail('Should render a button');
+			return fail('Should have role="button"');
 		}
 		cleanup(container);
 		pass();
@@ -46,7 +46,7 @@ export default {
 		await control.updateComplete;
 		let called = false;
 		editor.copyToClipboard = () => { called = true; return editor; };
-		control.shadowRoot.querySelector('button').click();
+		control.click();
 		if(!called){
 			cleanup(container);
 			return fail('Should call editor.copyToClipboard() on click');